BENCH-182 Implemented wrapper types over primitive

+++ src/main/java/com/answerdigital/answerking/request/LineItemRequest.java
@@ -5,7 +5,7 @@ import javax.validation.constraints.PositiveOrZero;
 
 public record LineItemRequest(
         @Positive(message = "The product ID must be a positive number")
-        int productId,
+        Long productId,
         @PositiveOrZero(message = "The quantity must be must be greater than or equal to 0")
-        int quantity
+        Integer quantity
 ) { }
